Correct timezone err on ‘is active’ slot indicator
The app relied on the client to determine “now” - what the time currently is.

The Slot component then used this timestamp to check if a current slot in the schedule was active.

I added a well supported timezone library to anchor the “now” timestamp to the “Atlantic/Reykjavik” timezone so the “is-active” indicator always accurately reflects the conference schedule.

Fixes #100